pyproject.toml
[tool.poetry]
name = "prognet"
version = "0.1.0"
description = ""
authors = ["Timofey Danshin <t.danshin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.6.2"
aiopg = {extras = ["sa"], version = "^1.0.0"}
aiohttp-jinja2 = "^1.2.0"
trafaret-config = "^2.0.2"
pyyaml = "^5.3.1"
aiohttp-security = "^0.4.0"
aiohttp-session = "^2.9.0"
isort = "^5.0.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
